Moves "requestHandlerUtils" into "utils/handlers" directory
